Just committing stuff that was on my laptop
and not yet committed.  None of this is finished.  I think I
know how to do both LRT and Rao tests, but haven't finished
the design doc or R function anova.glmdr.  So all of these
are broken at this point.  They are just there so Dan can
see what I have been working on.